+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Fixed `TreeStore` - 2nd arg `action` can be undefined (PR #1054) (issue #1056)
   - Fixed import of negative number in SpEL (PR #1054) (issue #1055)
   - Fixed validation error message for min/max (PR #1058) (issue #1057)
+  - (Ternary mode) Added to `config.settings`: `caseValueField`, `renderBeforeCaseValue`, `renderAfterCaseValue`, `renderSwitchPrefix`, `defaultCaseLabel`, `addCaseLabel`, `addDefaultCaseLabel` (PR #1062)
 - 6.5.2
   - Updated dependencies. `@babel/runtime` is now dep for core package (PR #1051) (issue #964)
 - 6.5.1